Cleaned up code and small styling
