+++ src/components/AboutUs/AboutUs.css
@@ -1,0 +1,3 @@
+div.aboutUs{
+    height: 100vh;
+}
