+++ src/components/App.js
@@ -1,18 +1,11 @@
 import React, { Component } from 'react';
-import Header from './Header/Header.js';
 import home from './home.js';
-import AboutUs from './AboutUs.js';
-import Footer from './Footer/Footer.js';
-import WeatherData from './WeatherCard/WeatherData.js';
-import WeatherCard from './WeatherCard/WeatherCard.js';
-import TrafficCard from './TrafficCard/TrafficCard.js';
-import TrafficCardArchive from './TrafficCardArchive.js';
-import WeatherCardArchive from './WeatherCardArchive.js';
-import CardContainer from './CardContainer/CardContainer.js';
+import AboutUs from './AboutUs/AboutUs.js';
+import TrafficCardArchive from './CardArchive/TrafficCardArchive.js';
+import WeatherCardArchive from './CardArchive/WeatherCardArchive.js';
 import {
   BrowserRouter as Router,
-  Route,
-  Link
+  Route
 } from 'react-router-dom'
 
 import './App.css';
