+++ src/components/Header/Header.js
@@ -1,10 +1,5 @@
 import React, { Component } from 'react'
 import { Menu, Dropdown } from 'semantic-ui-react'
-import {
-    BrowserRouter as Router,
-    Route,
-    Link
-} from 'react-router-dom'
 
 import './Header.css';
