+++ src/components/home.js
@@ -2,15 +2,8 @@ import React, { Component } from 'react';
 import Header from './Header/Header.js';
 import Footer from './Footer/Footer.js';
 import WeatherData from './WeatherCard/WeatherData.js';
-import WeatherCardArchive from './WeatherCardArchive.js';
-import TrafficCardArchive from './TrafficCardArchive.js';
 import TrafficCard from './TrafficCard/TrafficCard.js';
 import CardContainer from './CardContainer/CardContainer.js';
-import {
-  BrowserRouter as Router,
-  Route,
-  Link
-} from 'react-router-dom'
 
 import './App.css';
 
@@ -37,7 +30,9 @@ export default class home extends Component {
             label: card.props.menuLabel,
             index: index
         });
+
         return (
+        <html class="home-bg">
             <div className='home'>
                 <Header
                     cardEntries={cards.map(cardToDropdown)}
@@ -53,6 +48,7 @@ export default class home extends Component {
                 </div>
                 <Footer />
             </div>
+            </html>
         );
     }
 }
