+++ src/index.css
@@ -1,9 +1,19 @@
 body {
-  border-radius: 25px;
-  margin: 0;
-  padding: 0;
-  font-family: Lato,'Helvetica Neue',Arial,Helvetica,sans-serif;
-  background:
-        url('./images/blurred-background-5.jpg');
-  background-size: cover;
+    border-radius: 25px;
+    margin: 0;
+    padding: 0;
+    font-family: Lato,'Helvetica Neue',Arial,Helvetica,sans-serif;
+    background: url('./images/blurred-background-5.jpg');
+    background-size: cover;
+}
+
+html {
+    -webkit-background-size: cover;
+    -moz-background-size: cover;
+    -o-background-size: cover;
+    background-size: cover;
+}
+
+.aboutUs-bg {
+    background: white;
 }
